Made it more reusable
