+++ mulighetsrommet-api/src/test/kotlin/no/nav/mulighetsrommet/api/utbetaling/ArrangorflateRoutesTest.kt
@@ -9,151 +9,46 @@ import io.ktor.client.engine.mock.*
 import io.ktor.client.plugins.contentnegotiation.*
 import io.ktor.client.request.*
 import io.ktor.http.*
-import io.ktor.http.content.*
 import io.ktor.serialization.kotlinx.json.*
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonElement
 import kotliquery.Query
 import no.nav.amt.model.EndringAarsak
 import no.nav.amt.model.Melding
-import no.nav.mulighetsrommet.altinn.AltinnClient
-import no.nav.mulighetsrommet.altinn.AltinnClient.AuthorizedParty
-import no.nav.mulighetsrommet.altinn.AltinnClient.AuthorizedPartyType
 import no.nav.mulighetsrommet.api.arrangor.model.ArrangorDto
 import no.nav.mulighetsrommet.api.arrangorflate.GodkjennUtbetaling
 import no.nav.mulighetsrommet.api.arrangorflate.model.ArrFlateUtbetaling
-import no.nav.mulighetsrommet.api.clients.dokark.DokarkResponse
-import no.nav.mulighetsrommet.api.clients.dokark.DokarkResponseDokument
-import no.nav.mulighetsrommet.api.createAuthConfig
-import no.nav.mulighetsrommet.api.createTestApplicationConfig
 import no.nav.mulighetsrommet.api.databaseConfig
-import no.nav.mulighetsrommet.api.fixtures.*
-import no.nav.mulighetsrommet.api.tilsagn.db.TilsagnDbo
 import no.nav.mulighetsrommet.api.tilsagn.model.ArrangorflateTilsagn
-import no.nav.mulighetsrommet.api.tilsagn.model.TilsagnBeregningFri
 import no.nav.mulighetsrommet.api.tilsagn.model.TilsagnStatus
-import no.nav.mulighetsrommet.api.tilsagn.model.TilsagnType
-import no.nav.mulighetsrommet.api.utbetaling.db.DeltakerDbo
 import no.nav.mulighetsrommet.api.utbetaling.db.DeltakerForslag
 import no.nav.mulighetsrommet.api.utbetaling.db.DeltakerForslag.Status
-import no.nav.mulighetsrommet.api.utbetaling.db.UtbetalingDbo
-import no.nav.mulighetsrommet.api.utbetaling.model.DeltakelseManedsverk
-import no.nav.mulighetsrommet.api.utbetaling.model.DeltakelsePeriode
-import no.nav.mulighetsrommet.api.utbetaling.model.DeltakelsePerioder
-import no.nav.mulighetsrommet.api.utbetaling.model.UtbetalingBeregningForhandsgodkjent
 import no.nav.mulighetsrommet.api.withTestApplication
 import no.nav.mulighetsrommet.database.kotest.extensions.ApiDatabaseTestListener
-import no.nav.mulighetsrommet.ktor.MockEngineBuilder
 import no.nav.mulighetsrommet.ktor.createMockEngine
-import no.nav.mulighetsrommet.ktor.respondJson
-import no.nav.mulighetsrommet.model.DeltakerStatus
 import no.nav.mulighetsrommet.model.Kontonummer
-import no.nav.mulighetsrommet.model.NorskIdent
-import no.nav.mulighetsrommet.model.Periode
 import no.nav.security.mock.oauth2.MockOAuth2Server
-import java.time.LocalDate
-import java.time.LocalDateTime
 import java.util.*
 
 class ArrangorflateRoutesTest : FunSpec({
     val database = extension(ApiDatabaseTestListener(databaseConfig))
-
-    val identMedTilgang = NorskIdent("01010199988")
-
-    val hovedenhet = ArrangorFixtures.hovedenhet
-    val underenhet = ArrangorFixtures.underenhet1
-
-    val deltaker = DeltakerDbo(
-        id = UUID.randomUUID(),
-        gjennomforingId = GjennomforingFixtures.AFT1.id,
-        startDato = GjennomforingFixtures.AFT1.startDato,
-        sluttDato = GjennomforingFixtures.AFT1.sluttDato,
-        registrertTidspunkt = GjennomforingFixtures.AFT1.startDato.atStartOfDay(),
-        endretTidspunkt = LocalDateTime.now(),
-        deltakelsesprosent = 100.0,
-        deltakelsesmengder = listOf(),
-        status = DeltakerStatus(
-            type = DeltakerStatus.Type.DELTAR,
-            aarsak = null,
-            opprettetDato = LocalDateTime.now(),
-        ),
-    )
-
-    val tilsagn = TilsagnDbo(
-        id = UUID.randomUUID(),
-        gjennomforingId = GjennomforingFixtures.AFT1.id,
-        periode = Periode(LocalDate.of(2024, 6, 1), LocalDate.of(2025, 1, 1)),
-        lopenummer = 1,
-        bestillingsnummer = "A-2025/1-1",
-        kostnadssted = NavEnhetFixtures.Innlandet.enhetsnummer,
-        beregning = TilsagnBeregningFri(
-            input = TilsagnBeregningFri.Input(1000),
-            output = TilsagnBeregningFri.Output(1000),
-        ),
-        arrangorId = ArrangorFixtures.underenhet1.id,
-        type = TilsagnType.TILSAGN,
-    )
-
-    val utbetaling = UtbetalingDbo(
-        id = UUID.randomUUID(),
-        gjennomforingId = GjennomforingFixtures.AFT1.id,
-        fristForGodkjenning = LocalDateTime.now(),
-        beregning = UtbetalingBeregningForhandsgodkjent(
-            input = UtbetalingBeregningForhandsgodkjent.Input(
-                periode = Periode.forMonthOf(LocalDate.of(2024, 8, 1)),
-                sats = 20205,
-                stengt = setOf(),
-                deltakelser = setOf(
-                    DeltakelsePerioder(
-                        deltakelseId = deltaker.id,
-                        perioder = listOf(
-                            DeltakelsePeriode(
-                                periode = Periode(LocalDate.of(2024, 8, 1), LocalDate.of(2024, 8, 31)),
-                                deltakelsesprosent = 100.0,
-                            ),
-                        ),
-                    ),
-                ),
-            ),
-            output = UtbetalingBeregningForhandsgodkjent.Output(
-                belop = 0,
-                deltakelser = setOf(
-                    DeltakelseManedsverk(
-                        deltakelseId = deltaker.id,
-                        manedsverk = 1.0,
-                    ),
-                ),
-            ),
-        ),
-        kontonummer = Kontonummer("12312312312"),
-        kid = null,
-        periode = Periode.forMonthOf(LocalDate.of(2024, 8, 1)),
-        innsender = null,
-        beskrivelse = null,
-    )
-
-    val domain = MulighetsrommetTestDomain(
-        navEnheter = listOf(NavEnhetFixtures.IT, NavEnhetFixtures.Innlandet, NavEnhetFixtures.Gjovik),
-        ansatte = listOf(NavAnsattFixture.ansatt1, NavAnsattFixture.ansatt2),
-        tiltakstyper = listOf(TiltakstypeFixtures.AFT),
-        avtaler = listOf(
-            AvtaleFixtures.AFT.copy(
-                arrangor = AvtaleFixtures.AFT.arrangor?.copy(
-                    hovedenhet = hovedenhet.id,
-                    underenheter = listOf(underenhet.id),
-                ),
-            ),
-        ),
-        gjennomforinger = listOf(GjennomforingFixtures.AFT1.copy(arrangorId = underenhet.id)),
-        deltakere = listOf(deltaker),
-        arrangorer = listOf(hovedenhet, underenhet),
-        tilsagn = listOf(tilsagn),
-        utbetalinger = listOf(utbetaling),
-    ) {
-        setTilsagnStatus(tilsagn, TilsagnStatus.GODKJENT)
-    }
-
     val oauth = MockOAuth2Server()
+
+    // Use test utils to create test data
+    val deltaker = ArrangorflateTestUtils.createTestDeltaker()
+    val tilsagn = ArrangorflateTestUtils.createTestTilsagn()
+    val utbetaling = ArrangorflateTestUtils.createTestUtbetalingForhandsgodkjent(deltaker.id)
+
+    // Set up domain with test data
+    val domain = ArrangorflateTestUtils.createTestDomain(
+        deltaker = deltaker,
+        tilsagn = tilsagn,
+        utbetalinger = listOf(utbetaling),
+    )
+
+    val identMedTilgang = ArrangorflateTestUtils.identMedTilgang
+    val underenhet = ArrangorflateTestUtils.underenhet
+    val orgnr = underenhet.organisasjonsnummer.value
 
     beforeSpec {
         oauth.start()
@@ -171,56 +66,8 @@ class ArrangorflateRoutesTest : FunSpec({
         oauth.shutdown()
     }
 
-    fun MockEngineBuilder.mockAltinnAuthorizedParties() {
-        post("/altinn/accessmanagement/api/v1/resourceowner/authorizedparties") {
-            val body = Json.decodeFromString<AltinnClient.AltinnRequest>(
-                (it.body as TextContent).text,
-            )
-            if (body.value == identMedTilgang.value) {
-                respondJson(
-                    listOf(
-                        AuthorizedParty(
-                            organizationNumber = underenhet.organisasjonsnummer.value,
-                            name = underenhet.navn,
-                            type = AuthorizedPartyType.Organization,
-                            authorizedResources = listOf("nav_tiltaksarrangor_be-om-utbetaling"),
-                            subunits = emptyList(),
-                        ),
-                    ),
-                )
-            } else {
-                respondJson(emptyList<AuthorizedParty>())
-            }
-        }
-    }
-
-    fun MockEngineBuilder.mockJournalpost() {
-        post("/dokark/rest/journalpostapi/v1/journalpost") {
-            respondJson(
-                DokarkResponse(
-                    journalpostId = "123",
-                    journalstatus = "bra",
-                    melding = null,
-                    journalpostferdigstilt = true,
-                    dokumenter = listOf(DokarkResponseDokument("123")),
-                ),
-            )
-        }
-    }
-
-    fun appConfig(
-        engine: MockEngine = createMockEngine {
-            mockAltinnAuthorizedParties()
-            mockJournalpost()
-        },
-    ) = createTestApplicationConfig().copy(
-        database = databaseConfig,
-        auth = createAuthConfig(oauth, roles = emptyList()),
-        engine = engine,
-    )
-
     test("401 Unauthorized uten pid i claims") {
-        withTestApplication(appConfig()) {
+        withTestApplication(ArrangorflateTestUtils.appConfig(oauth)) {
             val response = client.get("/api/v1/intern/arrangorflate/tilgang-arrangor") {
                 bearerAuth(oauth.issueToken().serialize())
             }
@@ -229,7 +76,7 @@ class ArrangorflateRoutesTest : FunSpec({
     }
 
     test("200 OK og tom liste med pid uten tilgang") {
-        withTestApplication(appConfig()) {
+        withTestApplication(ArrangorflateTestUtils.appConfig(oauth)) {
             val client = createClient {
                 install(ContentNegotiation) {
                     json()
@@ -246,7 +93,7 @@ class ArrangorflateRoutesTest : FunSpec({
     }
 
     test("200 og arrangor returneres på tilgang endepunkt") {
-        withTestApplication(appConfig()) {
+        withTestApplication(ArrangorflateTestUtils.appConfig(oauth)) {
             val client = createClient {
                 install(ContentNegotiation) {
                     json()
@@ -267,10 +114,8 @@ class ArrangorflateRoutesTest : FunSpec({
         }
     }
 
-    val orgnr = underenhet.organisasjonsnummer.value
-
     test("403 hent tilsagn uten tilgang til bedrift") {
-        withTestApplication(appConfig()) {
+        withTestApplication(ArrangorflateTestUtils.appConfig(oauth)) {
             val response = client.get("/api/v1/intern/arrangorflate/arrangor/$orgnr/tilsagn") {
                 bearerAuth(oauth.issueToken(claims = mapOf("pid" to "01010199922")).serialize())
                 contentType(ContentType.Application.Json)
@@ -280,7 +125,7 @@ class ArrangorflateRoutesTest : FunSpec({
     }
 
     test("200 hent tilsagn med tilgang til bedrift") {
-        withTestApplication(appConfig()) {
+        withTestApplication(ArrangorflateTestUtils.appConfig(oauth)) {
             val client = createClient {
                 install(ContentNegotiation) {
                     json()
@@ -304,7 +149,7 @@ class ArrangorflateRoutesTest : FunSpec({
     }
 
     test("403 hent utbetaling uten tilgang til bedrift") {
-        withTestApplication(appConfig()) {
+        withTestApplication(ArrangorflateTestUtils.appConfig(oauth)) {
             val response = client.get("/api/v1/intern/arrangorflate/utbetaling/${utbetaling.id}") {
                 bearerAuth(oauth.issueToken(claims = mapOf("pid" to "01010199922")).serialize())
                 contentType(ContentType.Application.Json)
@@ -314,7 +159,7 @@ class ArrangorflateRoutesTest : FunSpec({
     }
 
     test("200 hent utbetaling") {
-        withTestApplication(appConfig()) {
+        withTestApplication(ArrangorflateTestUtils.appConfig(oauth)) {
             val client = createClient {
                 install(ContentNegotiation) {
                     json()
@@ -335,7 +180,7 @@ class ArrangorflateRoutesTest : FunSpec({
     }
 
     test("400 ved feil sjekksum ved godkjenning av utbetaling") {
-        withTestApplication(appConfig()) {
+        withTestApplication(ArrangorflateTestUtils.appConfig(oauth)) {
             val client = createClient {
                 install(ContentNegotiation) {
                     json()
@@ -379,7 +224,7 @@ class ArrangorflateRoutesTest : FunSpec({
 
     // TODO: flytt resten av godkjenning-testene til egen testklasse for ArrangorflateService
     test("riktig sjekksum ved godkjenning av utbetaling gir 200, og spawner journalforing task") {
-        withTestApplication(appConfig()) {
+        withTestApplication(ArrangorflateTestUtils.appConfig(oauth)) {
             val client = createClient {
                 install(ContentNegotiation) {
                     json()
@@ -410,7 +255,7 @@ class ArrangorflateRoutesTest : FunSpec({
     }
 
     test("kan ikke godkjenne allerede godkjent") {
-        withTestApplication(appConfig()) {
+        withTestApplication(ArrangorflateTestUtils.appConfig(oauth)) {
             val client = createClient {
                 install(ContentNegotiation) {
                     json()
@@ -449,15 +294,19 @@ class ArrangorflateRoutesTest : FunSpec({
     }
 
     test("feil mot dokark gir fortsatt 200 på godkjenn siden det skjer i en task") {
+        // Create custom engine to test error scenario
         val clientEngine = createMockEngine {
-            mockAltinnAuthorizedParties()
+            ArrangorflateTestUtils.mockAltinnAuthorizedParties(this)
 
             post("/dokark/rest/journalpostapi/v1/journalpost") {
                 respondError(HttpStatusCode.InternalServerError)
             }
         }
 
-        withTestApplication(appConfig(clientEngine)) {
+        // Use custom app config with error-producing engine
+        val errorConfig = ArrangorflateTestUtils.appConfig(oauth, engine = clientEngine)
+
+        withTestApplication(errorConfig) {
             val client = createClient {
                 install(ContentNegotiation) {
                     json()
@@ -495,7 +344,7 @@ class ArrangorflateRoutesTest : FunSpec({
             )
         }
 
-        withTestApplication(appConfig()) {
+        withTestApplication(ArrangorflateTestUtils.appConfig(oauth)) {
             val client = createClient {
                 install(ContentNegotiation) {
                     json()
